+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Smart chunking in chunk text extension erroneously disabled
 - Invalid text modification when stream completion contains single line
+- Invalid text modification when stream completion adds text above visible range
 
 ## [2024.10.2] - 2024-09-09